Extract AOC Governor
